+++ graphdatascience/graph/graph_alpha_proc_runner.py
@@ -4,6 +4,7 @@ from typing import List, Optional, Union
 from pandas import DataFrame
 
 from ..error.client_only_endpoint import client_only_endpoint
+from ..error.deprecation_warning import deprecation_warning
 from ..error.illegal_attr_checker import IllegalAttrChecker
 from ..error.uncallable_namespace import UncallableNamespace
 from ..server_version.compatible_with import compatible_with
@@ -36,6 +37,7 @@ class GraphAlphaProcRunner(UncallableNamespace, IllegalAttrChecker):
         return GraphAlphaProjectRunner(self._query_runner, self._namespace, self._server_version)
 
     @client_only_endpoint("gds.alpha.graph")
+    @deprecation_warning("gds.graph.construct", ServerVersion(2, 1, 0))
     @compatible_with("construct", min_inclusive=ServerVersion(2, 1, 0))
     def construct(
         self,