+++ graphdatascience/graph/graph_proc_runner.py
@@ -1,3 +1,4 @@
+import os
 import pathlib
 import sys
 from logging import warning
@@ -13,7 +14,6 @@ from ..error.illegal_attr_checker import IllegalAttrChecker
 from ..error.uncallable_namespace import UncallableNamespace
 from ..server_version.compatible_with import compatible_with
 from ..server_version.server_version import ServerVersion
-from .graph_alpha_proc_runner import GraphAlphaProcRunner
 from .graph_entity_ops_runner import (
     GraphElementPropertyRunner,
     GraphNodePropertiesRunner,
@@ -45,6 +45,53 @@ class GraphProcRunner(UncallableNamespace, IllegalAttrChecker):
             return path(package, resource)
 
     @client_only_endpoint("gds.graph")
+    @compatible_with("construct", min_inclusive=ServerVersion(2, 1, 0))
+    def construct(
+        self,
+        graph_name: str,
+        nodes: Union[DataFrame, List[DataFrame]],
+        relationships: Union[DataFrame, List[DataFrame]],
+        concurrency: int = 4,
+        undirected_relationship_types: Optional[List[str]] = None,
+    ) -> Graph:
+        nodes = nodes if isinstance(nodes, List) else [nodes]
+        relationships = relationships if isinstance(relationships, List) else [relationships]
+
+        errors = []
+
+        exists = self._query_runner.run_query(
+            f"CALL gds.graph.exists('{graph_name}') YIELD exists", custom_error=False
+        ).squeeze()
+
+        # compare against True as (1) unit tests return None here and (2) numpys True does not work with `is True`.
+        if exists == True:  # noqa: E712
+            errors.append(
+                f"Graph '{graph_name}' already exists. Please drop the existing graph or use a different name."
+            )
+
+        for idx, node_df in enumerate(nodes):
+            if "nodeId" not in node_df.columns.values:
+                errors.append(f"Node dataframe at index {idx} needs to contain a 'nodeId' column.")
+
+        for idx, rel_df in enumerate(relationships):
+            for expected_col in ["sourceNodeId", "targetNodeId"]:
+                if expected_col not in rel_df.columns.values:
+                    errors.append(f"Relationship dataframe at index {idx} needs to contain a '{expected_col}' column.")
+
+        if self._server_version < ServerVersion(2, 3, 0) and undirected_relationship_types:
+            errors.append("The parameter 'undirected_relationship_types' is only supported since GDS 2.3.0.")
+
+        if len(errors) > 0:
+            raise ValueError(os.linesep.join(errors))
+
+        constructor = self._query_runner.create_graph_constructor(
+            graph_name, concurrency, undirected_relationship_types
+        )
+        constructor.run(nodes, relationships)
+
+        return Graph(graph_name, self._query_runner, self._server_version)
+
+    @client_only_endpoint("gds.graph")
     def load_cora(self, graph_name: str = "cora", undirected: bool = False) -> Graph:
         with self._path("graphdatascience.resources.cora", "cora_nodes_gzip.pkl") as nodes_resource:
             nodes = read_pickle(nodes_resource, compression="gzip")
@@ -52,14 +99,9 @@ class GraphProcRunner(UncallableNamespace, IllegalAttrChecker):
         with self._path("graphdatascience.resources.cora", "cora_rels_gzip.pkl") as rels_resource:
             rels = read_pickle(rels_resource, compression="gzip")
 
-        self._namespace = "gds.alpha.graph"
-        alpha_proc_runner = GraphAlphaProcRunner(self._query_runner, self._namespace, self._server_version)
-
         undirected_relationship_types = ["*"] if undirected else []
 
-        return alpha_proc_runner.construct(
-            graph_name, nodes, rels, undirected_relationship_types=undirected_relationship_types
-        )
+        return self.construct(graph_name, nodes, rels, undirected_relationship_types=undirected_relationship_types)
 
     @client_only_endpoint("gds.graph")
     def load_karate_club(self, graph_name: str = "karate_club", undirected: bool = False) -> Graph:
@@ -69,14 +111,9 @@ class GraphProcRunner(UncallableNamespace, IllegalAttrChecker):
         with self._path("graphdatascience.resources.karate", "karate_club_gzip.pkl") as rels_resource:
             rels = read_pickle(rels_resource, compression="gzip")
 
-        self._namespace = "gds.alpha.graph"
-        alpha_proc_runner = GraphAlphaProcRunner(self._query_runner, self._namespace, self._server_version)
-
         undirected_relationship_types = ["*"] if undirected else []
 
-        return alpha_proc_runner.construct(
-            graph_name, nodes, rels, undirected_relationship_types=undirected_relationship_types
-        )
+        return self.construct(graph_name, nodes, rels, undirected_relationship_types=undirected_relationship_types)
 
     @client_only_endpoint("gds.graph")
     def load_imdb(self, graph_name: str = "imdb", undirected: bool = True) -> Graph:
@@ -97,18 +134,13 @@ class GraphProcRunner(UncallableNamespace, IllegalAttrChecker):
         with self._path("graphdatascience.resources.imdb", "imdb_directed_in_rels_gzip.pkl") as rels_resource:
             directed_in_rels = read_pickle(rels_resource, compression="gzip")
 
-        self._namespace = "gds.alpha.graph"
-        alpha_proc_runner = GraphAlphaProcRunner(self._query_runner, self._namespace, self._server_version)
-
         nodes = [movies_with_genre, movies_without_genre, actors, directors]
         rels = [acted_in_rels, directed_in_rels]
 
         # Default undirected which matches raw data
         undirected_relationship_types = ["*"] if undirected else []
 
-        return alpha_proc_runner.construct(
-            graph_name, nodes, rels, undirected_relationship_types=undirected_relationship_types
-        )
+        return self.construct(graph_name, nodes, rels, undirected_relationship_types=undirected_relationship_types)
 
     @property
     def sample(self) -> GraphSampleRunner: