+++ changelog.md
@@ -13,6 +13,7 @@
 	* `gds.graph.ogbl.load` for link property prediction datasets.
 * Added possibility to load NetworkX graphs via the new method `gds.graph.networkx.load`.
 * Added new sphinx API reference documentation for all procedures.
+* Promoted `gds.alpha.graph.construct` to `gds.graph.construct`.
 
 ## Bug fixes